shifted create_or_update into utils. regenerated test fixtures.
