+++ src/observer/utils.py
@@ -6,7 +6,14 @@ from datetime import datetime
 import pytz
 import itertools
 import logging
+
 LOG = logging.getLogger(__name__)
+
+def ensure(assertion, msg, *args):
+    """intended as a convenient replacement for `assert` statements that
+    get compiled away with -O flags"""
+    if not assertion:
+        raise AssertionError(msg % args)
 
 def listfiles(path, ext_list=None):
     "returns a list of absolute paths for given dir"
@@ -109,3 +116,32 @@ def take(n, items):
 
 def pad_msid(msid):
     return '%05d' % int(msid)
+
+EXCLUDE_ME = 0xDEADBEEF
+
+def create_or_update(Model, orig_data, key_list, create=True, update=True, commit=True, **overrides):
+    inst = None
+    created = updated = False
+    data = {}
+    data.update(orig_data)
+    data.update(overrides)
+    try:
+        # try and find an entry of Model using the key fields in the given data
+        inst = Model.objects.get(**subdict(data, key_list))
+        # object exists, otherwise DoesNotExist would have been raised
+        if update:
+            [setattr(inst, key, val) for key, val in data.items() if val != EXCLUDE_ME]
+            updated = True
+    except Model.DoesNotExist:
+        if create:
+            #inst = Model(**data)
+            # shift this exclude me handling to et3
+            inst = Model(**{k: v for k, v in data.items() if v != EXCLUDE_ME})
+            created = True
+
+    if (updated or created) and commit:
+        inst.save()
+
+    # it is possible to neither create nor update.
+    # in this case if the model cannot be found then None is returned: (None, False, False)
+    return (inst, created, updated)
